refacting UserInterface and MovementHandler
