+++ backend/src/main/java/chess/vmsb/logic/business/utils/MovementHandler.java
@@ -1,6 +1,7 @@
 package chess.vmsb.logic.business.utils;
 
 import chess.vmsb.logic.business.board.Board;
+import chess.vmsb.logic.business.board.Square;
 import chess.vmsb.logic.business.models.Piece;
 import chess.vmsb.logic.business.models.Player;
 
@@ -34,8 +35,37 @@ public class MovementHandler implements Functional{
         }
         return wbKingXY;
     }
-    protected boolean isValidMove(Board board, ArrayList<ArrayList<Integer>> moveData){
-        //TODO implement isValidMove logic
+    public boolean isValidMove(Board board, ArrayList<ArrayList<Integer>> moveData){
+        // Extract the 'from' and 'to' coordinates from the moveData
+        ArrayList<Integer> from = moveData.get(0);
+        ArrayList<Integer> to = moveData.get(1);
+        // Get the file (column) and rank (row) values from the 'from' and 'to' coordinates
+        int fromFile = from.get(0);
+        int fromRank = from.get(1);
+        int toFile = to.get(0);
+        int toRank = to.get(1);
+        // Validate if the 'from' and 'to' coordinates are within the valid board range
+        if (fromFile < 0 || fromFile > 7 || fromRank < 0 || fromRank > 7 || toFile < 0 || toFile > 7 || toRank < 0
+                || toRank > 7) {
+            System.out.println("Invalid move. Coordinates are out of range.");
+            return false;
+        }
+        // Get the square at the 'from' and 'to' coordinates on the board
+        Square fromSquare = board.getGameBoard()[fromRank][fromFile];
+        Square toSquare = board.getGameBoard()[toRank][toFile];
+        // Validate if the 'from' square does not contains a piece and the 'to' square is empty
+        if (fromSquare.getPiece() == null || toSquare.getPiece() != null) {
+            System.out.println("Invalid move. Source square does not contain a piece or target square is not empty.");
+            return false;
+        } else if(fromSquare.getPiece().isMoved() == toSquare.getPiece().isMoved()){
+            System.out.println("Invalid move. You can't move your piece to a square occupied by your own piece.");
+            return false;
+        }
+        // [Add additional logic to check if the move is valid based on your game rules]
+        // For example, you can check if the move is valid for the specific piece type,
+        // the player's turn, the direction of the move, etc.
+
+        // If all validations pass, return true to indicate a valid move
         return true;
     }
     //# performMove(Board board, Player[] player, ArrayList<ArrayList<Integer>> moveData): Object[]
